+++ python/src/marginpy/generated_client/accounts/__init__.py
@@ -1,4 +1,3 @@
-# flake8: noqa
 from .marginfi_account import MarginfiAccount, MarginfiAccountJSON
 from .marginfi_group import MarginfiGroup, MarginfiGroupJSON
 from .state import State, StateJSON