+++ python/src/marginpy/generated_client/instructions/__init__.py
@@ -1,4 +1,3 @@
-# flake8: noqa
 from .bank_fee_vault_withdraw import (
     BankFeeVaultWithdrawAccounts,
     BankFeeVaultWithdrawArgs,